Tidy up #1
