+++ src/frontend/packages/core/src/features/endpoints/backup-restore/backup-restore.types.ts
@@ -1,0 +1,25 @@
+import { EndpointModel } from '../../../../../store/src/types/endpoint.types';
+
+export enum BackupEndpointTypes {
+  ENDPOINT = 'endpoint',
+  CONNECT = 'connect',
+}
+
+export enum BackupEndpointConnectionTypes {
+  NONE = 'NONE',
+  CURRENT = 'CURRENT',
+  ALL = 'ALL'
+}
+
+export interface BackupEndpointsConfig<T> {
+  [endpointId: string]: T;
+}
+
+export interface BaseEndpointConfig {
+  [BackupEndpointTypes.ENDPOINT]: boolean;
+  [BackupEndpointTypes.CONNECT]: BackupEndpointConnectionTypes;
+}
+
+export interface BackupEndpointConfigUI extends BaseEndpointConfig {
+  entity: EndpointModel;
+}
