+++ src/frontend/packages/core/src/jetstream.helpers.ts
@@ -53,3 +53,20 @@ export function isHttpErrorResponse(obj: any): HttpErrorResponse {
   ) ? obj as HttpErrorResponse : null;
 }
 
+/**
+ * Attempt to create a sensible string explaining the error object returned from a failed http request
+ * @param err The raw error from a http request
+ */
+export function httpErrorResponseToSafeString(err: any): string {
+  const httpResponse: HttpErrorResponse = isHttpErrorResponse(err);
+  if (httpResponse) {
+    if (httpResponse.error) {
+      if (typeof (httpResponse.error) === 'string') {
+        return httpResponse.error + ` (${httpResponse.status})`;
+      }
+      return httpResponse.error.error + ` (${httpResponse.status})`;
+    }
+    return JSON.stringify(httpResponse.error) + ` (${httpResponse.status})`;
+  }
+  return err.message;
+}
