+++ src/jetstream/cnsi.go
@@ -3,9 +3,7 @@ package main
 import (
 	"crypto/x509"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"strconv"
@@ -638,340 +636,24 @@ func (p *portalProxy) updateEndpoint(c echo.Context) error {
 	return nil
 }
 
-type BackupConnectionType string
-
-const (
-	BACKUP_CONNECTION_NONE    BackupConnectionType = "NONE"
-	BACKUP_CONNECTION_CURRENT                      = "CURRENT"
-	BACKUP_CONNECTION_ALL                          = "ALL"
-)
-
-// TODO: RC position
-type BackupEndpointsState struct {
-	Endpoint bool                 `json:"endpoint"`
-	Connect  BackupConnectionType `json:"connect"`
-}
-
-// Sent to backup
-type BackupRequest struct {
-	State     map[string]BackupEndpointsState `json:"state"`
-	UserID    string                          `json:"userId"`
-	DBVersion string                          `json:"dbVersion"`
-	Password  string                          `json:"password"`
-}
-
-type BackupRequestEndpointsResponse struct {
-	Endpoints []*interfaces.CNSIRecord
-	Tokens    []interfaces.BackupTokenRecord
-}
-
-type BackupContent struct {
-	payload   BackupRequestEndpointsResponse `json:"payload"`
-	DBVersion int64                          `json:"dbVersion"`
-}
-
-type RestoreRequest struct {
-	// Data - Encrypted version of BackupContent
-	Data            string `json:"data"`
-	Password        string `json:"password"`
-	IgnoreDbVersion bool   `json:"ignoreDbVersion"`
-}
-
-// TODO: RC split out to new file?
 func (p *portalProxy) backupEndpoints(c echo.Context) error {
-	log.Debug("backupEndpoints")
+	log.Debug("BackupEndpoints")
 
-	// Check we can unmarshall the request
-	body, err := ioutil.ReadAll(c.Request().Body)
-	if err != nil {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid request body")
+	ctb := &cnsiTokenBackup{
+		databaseConnectionPool: p.DatabaseConnectionPool,
+		p:                      p,
 	}
 
-	data := &BackupRequest{}
-	if err = json.Unmarshal(body, data); err != nil {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid request body - could not parse JSON")
-	}
-	// log.Infof("BODY: %+v", data)
-
-	if data.State == nil || len(data.State) == 0 {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid request body - no endpoints to backup")
-	}
-
-	response, err := p.createBackup(data)
-	if err != nil {
-		return err
-	}
-
-	log.Infof("response: %+v", response)
-
-	// Send back the response to the client
-	// TODO: RC Missing client_secret when serialised, `-` in definition
-	jsonString, err := json.Marshal(response)
-	if err != nil {
-		return interfaces.NewHTTPError(http.StatusInternalServerError, "Failed to serialize response")
-	}
-
-	log.Infof("jsonString: %+v", jsonString)
-
-	// Return data
-	c.Response().Header().Set("Content-Type", "application/json")
-	c.Response().Write(jsonString)
-	return nil
-}
-
-func (p *portalProxy) createBackup(data *BackupRequest) (*BackupContent, error) {
-	log.Debug("createBackup")
-	allEndpoints, err := p.ListEndpoints()
-	if err != nil {
-		return nil, interfaces.NewHTTPError(http.StatusBadGateway, "Failed to fetch endpoints")
-	}
-
-	// Fetch/Format required data
-	endpoints := make([]*interfaces.CNSIRecord, 0)
-	// allTokensFrom := make([]string, 0)
-	// userTokenFrom := make([]string, 0)
-	tokens := make([]interfaces.BackupTokenRecord, 0)
-
-	for endpointID, endpoint := range data.State {
-
-		if !endpoint.Endpoint {
-			continue
-		}
-
-		for _, e := range allEndpoints {
-			if endpointID == e.GUID {
-				endpoints = append(endpoints, e)
-				break
-			}
-		}
-
-		switch connectionType := endpoint.Connect; connectionType {
-		case BACKUP_CONNECTION_ALL:
-			// allTokensFrom = append(allTokensFrom, endpointID)
-			if tokenRecords, ok := p.getCNSITokenRecordsBackup(endpointID); ok {
-				log.Warn("tokens for AllConnect")
-				tokens = append(tokens, tokenRecords...)
-			} else {
-				log.Warn("No tokens for AllConnect")
-				// TODO: RC
-			}
-		case BACKUP_CONNECTION_CURRENT:
-			// userTokenFrom = append(userTokenFrom, endpointID)
-			if tokenRecord, ok := p.GetCNSITokenRecordWithDisconnected(endpointID, data.UserID); ok {
-				log.Warn("tokens for Connect")
-				// var btr BackupTokenRecord
-				// TODO: RC Q This will be the linked token as if it were the users token
-				var btr = interfaces.BackupTokenRecord{
-					// tokenRecord: tokenRecord,
-					TokenRecord:  tokenRecord,
-					EndpointGUID: endpointID,
-					TokenType:    "CNSI",
-					UserGUID:     data.UserID,
-				}
-
-				tokens = append(tokens, btr)
-			} else {
-				log.Warnf("No tokens for Connect: %+v,%+v", endpointID, data.UserID)
-				// TODO: RC
-				// msg := "Unable to retrieve CNSI token record."
-				// log.Debug(msg)
-				// return nil, nil, false
-			}
-		}
-
-		// if endpoint.Connect == BACKUP_CONNECTION_ALL {
-		// 	// allTokensFrom = append(allTokensFrom, endpointID)
-		// 	if tokenRecords, ok := p.getCNSITokenRecordsBackup(endpointID); ok {
-		// 		log.Warn("tokens for AllConnect")
-		// 		tokens = append(tokens, tokenRecords...)
-		// 	} else {
-		// 		log.Warn("No tokens for AllConnect")
-		// 		// TODO: RC
-		// 	}
-		// } else if endpoint.Connect {
-		// 	// userTokenFrom = append(userTokenFrom, endpointID)
-		// 	if tokenRecord, ok := p.GetCNSITokenRecordWithDisconnected(endpointID, data.UserID); ok {
-		// 		log.Warn("tokens for Connect")
-		// 		// var btr BackupTokenRecord
-		// 		// TODO: RC Q This will be the linked token as if it were the users token
-		// 		var btr = interfaces.BackupTokenRecord{
-		// 			// tokenRecord: tokenRecord,
-		// 			TokenRecord:  tokenRecord,
-		// 			EndpointGUID: endpointID,
-		// 			TokenType:    "CNSI",
-		// 			UserGUID:     data.UserID,
-		// 		}
-
-		// 		tokens = append(tokens, btr)
-		// 	} else {
-		// 		log.Warnf("No tokens for Connect: %+v,%+v", endpointID, data.UserID)
-		// 		// TODO: RC
-		// 		// msg := "Unable to retrieve CNSI token record."
-		// 		// log.Debug(msg)
-		// 		// return nil, nil, false
-		// 	}
-		// }
-	}
-
-	log.Infof("endpoints: %+v", endpoints)
-	// log.Infof("allTokensFrom: %+v", allTokensFrom)
-	// log.Infof("userTokenFrom: %+v", userTokenFrom)
-	log.Infof("tokens: %+v", tokens)
-
-	payload := &BackupRequestEndpointsResponse{
-		Endpoints: endpoints,
-		Tokens:    tokens,
-	}
-
-	// Encrypt data (see above) // TODO: RC leave until last
-	// encryptedPayload := payload
-
-	versions, err := p.getVersionsData()
-	if err != nil {
-		return nil, errors.New("Could not find database version")
-	}
-
-	// log.Infof("payload: %+v", payload)
-
-	response := &BackupContent{
-		payload:   *payload,
-		DBVersion: versions.DatabaseVersion,
-	}
-
-	// log.Infof("response: %+v", response)
-
-	return response, nil
-}
-
-// func (p *portalProxy) GetCNSITokens(cnsiGUID string) ([]interfaces.TokenRecord, bool) {
-// 	log.Debug("GetCNSITokens")
-// 	tokenRepo, err := tokens.NewPgsqlTokenRepository(p.DatabaseConnectionPool)
-// 	if err != nil {
-// 		return make([]interfaces.TokenRecord, 0), false
-// 	}
-
-// 	trs, err := tokenRepo.FindAllCNSITokenIncludeDisconnected(cnsiGUID, p.Config.EncryptionKeyInBytes)
-// 	if err != nil {
-// 		return make([]interfaces.TokenRecord, 0), false
-// 	}
-
-// 	return trs, true
-// }
-
-func (p *portalProxy) getCNSITokenRecordsBackup(endpointID string) ([]interfaces.BackupTokenRecord, bool) {
-	log.Debug("getCNSITokenRecordsBackup")
-	tokenRepo, err := tokens.NewPgsqlTokenRepository(p.DatabaseConnectionPool)
-	if err != nil {
-		return make([]interfaces.BackupTokenRecord, 0), false
-	}
-
-	trs, err := tokenRepo.FindAllCNSITokenBackup(endpointID, p.Config.EncryptionKeyInBytes)
-	if err != nil {
-		return make([]interfaces.BackupTokenRecord, 0), false
-	}
-
-	return trs, true
+	return ctb.BackupEndpoints(c)
 }
 
 func (p *portalProxy) restoreEndpoints(c echo.Context) error {
 	log.Debug("restoreEndpoints")
 
-	// Check we can unmarshall the request
-	body, err := ioutil.ReadAll(c.Request().Body)
-	if err != nil {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid request body")
+	ctb := &cnsiTokenBackup{
+		databaseConnectionPool: p.DatabaseConnectionPool,
+		p:                      p,
 	}
 
-	data := &RestoreRequest{}
-	if err = json.Unmarshal(body, data); err != nil {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid request body - could not parse JSON")
-	}
-
-	err = p.restoreBackup(data)
-	if err != nil {
-		// TODO: RC write error?
-		return err
-	}
-
-	// log.Warnf("BACKUP DATA: %+v", backup)
-	c.Response().WriteHeader(http.StatusOK)
-	return nil
-
+	return ctb.RestoreEndpoints(c)
 }
-
-func (p *portalProxy) restoreBackup(backup *RestoreRequest) error {
-	log.Debug("restoreBackup")
-
-	// TODO: RC check return types of these functions... if we return shadow error
-
-	data := &BackupContent{}
-	if err := json.Unmarshal([]byte(backup.Data), backup); err != nil {
-		return interfaces.NewHTTPError(http.StatusBadRequest, "Invalid backup - could not parse JSON")
-	}
-
-	if backup.IgnoreDbVersion == false {
-		versions, err := p.getVersionsData()
-		if err != nil {
-			return errors.New("Could not find database version")
-		}
-
-		if versions.DatabaseVersion != data.DBVersion {
-			errorStr := fmt.Sprintf("Incompatible database versions. Expected %+v but got %+v", versions.DatabaseVersion, data.DBVersion)
-			return interfaces.NewHTTPShadowError(http.StatusBadRequest, errorStr, errorStr)
-		}
-	}
-
-	unencryptedBackup := data.payload
-
-	cnsiRepo, err := cnsis.NewPostgresCNSIRepository(p.DatabaseConnectionPool)
-	if err != nil {
-		return interfaces.NewHTTPShadowError(http.StatusInternalServerError, "Failed to connect to db", "Failed to connect to db: %+v", err)
-	}
-
-	for _, endpoint := range unencryptedBackup.Endpoints {
-		if err := cnsiRepo.Overwrite(*endpoint, p.Config.EncryptionKeyInBytes); err != nil {
-			return interfaces.NewHTTPShadowError(http.StatusInternalServerError, "Failed to overwrite endpoints", "Failed to overwrite endpoint: %+v", endpoint.Name)
-		}
-	}
-
-	tokenRepo, err := tokens.NewPgsqlTokenRepository(p.DatabaseConnectionPool)
-	if err != nil {
-		return interfaces.NewHTTPShadowError(http.StatusInternalServerError, "Failed to connect to db", "Failed to connect to db: %+v", err)
-	}
-
-	for _, tr := range unencryptedBackup.Tokens {
-		if err := tokenRepo.SaveCNSIToken(tr.EndpointGUID, tr.UserGUID, tr.TokenRecord, p.Config.EncryptionKeyInBytes); err != nil {
-			return interfaces.NewHTTPShadowError(http.StatusInternalServerError, "Failed to overwrite token", "Failed to overwrite token: %+v", tr.TokenRecord.TokenGUID)
-		}
-	}
-
-	return nil
-}
-
-// find := func(a interfaces.CNSIRecord) bool {
-// 	return endpointID == a.GUID
-// }
-
-// endpointPos := sliceContainsFn(find, allEndpoints)
-// if endpointPos >= 0 {
-// 	endpoints = append(endpoints, endpoints[endpointPos])
-// }
-
-// // TODO:RC pos
-// func sliceContains(what interface{}, where []interface{}) (idx int) {
-// 	for i, v := range where {
-// 		if v == what {
-// 			return i
-// 		}
-// 	}
-// 	return -1
-// }
-
-// func sliceContainsFn(is func(a interface{}) bool, where []interface{}) (idx int) {
-// 	for i, v := range where {
-// 		if is(v) {
-// 			return i
-// 		}
-// 	}
-// 	return -1
-// }
