+++ src/jetstream/repository/tokens/pgsql_tokens.go
@@ -43,15 +43,6 @@ var findCNSITokenConnected = `SELECT token_guid, auth_token, refresh_token, toke
 										FROM tokens
 										WHERE cnsi_guid = $1 AND (user_guid = $2 OR user_guid = $3) AND token_type = 'cnsi' AND disconnected = '0'`
 
-// TODO: RC
-var findAllCNSIToken2 = `SELECT token_guid, auth_token, refresh_token, token_expiry, disconnected, auth_type, meta_data, user_guid, linked_token
-										FROM tokens
-										WHERE cnsi_guid = $1 AND token_type = 'cnsi'`
-
-var findAllCNSITokenConnected = `SELECT token_guid, auth_token, refresh_token, token_expiry, disconnected, auth_type, meta_data, user_guid, linked_token
-										FROM tokens
-										WHERE cnsi_guid = $1 AND token_type = 'cnsi' AND disconnected = '0'`
-
 var findAllCNSIToken = `SELECT user_guid, token_guid, auth_token, refresh_token, token_expiry, disconnected, auth_type, meta_data, user_guid, linked_token
 										FROM tokens
 										WHERE cnsi_guid = $1 AND token_type = 'cnsi'`
@@ -95,15 +86,12 @@ func InitRepositoryProvider(databaseProvider string) {
 	updateAuthToken = datastore.ModifySQLStatement(updateAuthToken, databaseProvider)
 	findCNSIToken = datastore.ModifySQLStatement(findCNSIToken, databaseProvider)
 	findCNSITokenConnected = datastore.ModifySQLStatement(findCNSITokenConnected, databaseProvider)
-	findAllCNSIToken2 = datastore.ModifySQLStatement(findAllCNSIToken2, databaseProvider)
-	findAllCNSITokenConnected = datastore.ModifySQLStatement(findAllCNSITokenConnected, databaseProvider)
 	countCNSITokens = datastore.ModifySQLStatement(countCNSITokens, databaseProvider)
 	insertCNSIToken = datastore.ModifySQLStatement(insertCNSIToken, databaseProvider)
 	updateCNSIToken = datastore.ModifySQLStatement(updateCNSIToken, databaseProvider)
 	deleteCNSIToken = datastore.ModifySQLStatement(deleteCNSIToken, databaseProvider)
 	deleteCNSITokens = datastore.ModifySQLStatement(deleteCNSITokens, databaseProvider)
 	updateToken = datastore.ModifySQLStatement(updateToken, databaseProvider)
-	findAllCNSIToken = datastore.ModifySQLStatement(findAllCNSIToken, databaseProvider)
 }
 
 // saveAuthToken - Save the Auth token to the datastore
@@ -327,129 +315,6 @@ func (p *PgsqlTokenRepository) SaveCNSIToken(cnsiGUID string, userGUID string, t
 	}
 
 	return nil
-}
-
-// TODO: RC remove all `all`
-func (p *PgsqlTokenRepository) FindAllCNSIToken(cnsiGUID string, encryptionKey []byte) ([]interfaces.TokenRecord, error) {
-	log.Debug("FindAllCNSIToken")
-	return p.findAllCNSIToken(cnsiGUID, encryptionKey, false)
-}
-
-func (p *PgsqlTokenRepository) FindAllCNSITokenIncludeDisconnected(cnsiGUID string, encryptionKey []byte) ([]interfaces.TokenRecord, error) {
-	log.Debug("FindAllCNSITokenIncludeDisconnected")
-	return p.findAllCNSIToken(cnsiGUID, encryptionKey, true)
-}
-
-func (p *PgsqlTokenRepository) findAllCNSIToken(cnsiGUID string, encryptionKey []byte, includeDisconnected bool) ([]interfaces.TokenRecord, error) {
-	log.Debug("findAllCNSIToken")
-	if cnsiGUID == "" {
-		msg := "Unable to find CNSI Token without a valid CNSI GUID."
-		log.Debug(msg)
-		return make([]interfaces.TokenRecord, 0), errors.New(msg)
-	}
-
-	var rows *sql.Rows
-	var err error
-	if includeDisconnected {
-		rows, err = p.db.Query(findAllCNSIToken, cnsiGUID)
-	} else {
-		rows, err = p.db.Query(findAllCNSITokenConnected, cnsiGUID)
-	}
-	if err != nil {
-		msg := "Unable to Find All CNSI tokens: %v"
-		if err == sql.ErrNoRows {
-			log.Debugf(msg, err)
-		} else {
-			log.Errorf(msg, err)
-		}
-		return make([]interfaces.TokenRecord, 0), fmt.Errorf(msg, err)
-	}
-
-	// TODO: RC Q should this close come before returning? it doesn't in cnsi List(encryptionKey []byte) ([]*interfaces.CNSIRecord, error) {
-	defer rows.Close()
-
-	trs := make([]interfaces.TokenRecord, 0)
-	for rows.Next() {
-		// temp vars to retrieve db data
-		var (
-			tokenGUID              sql.NullString
-			ciphertextAuthToken    []byte
-			ciphertextRefreshToken []byte
-			tokenExpiry            sql.NullInt64
-			disconnected           bool
-			authType               string
-			metadata               sql.NullString
-			tokenUserGUID          sql.NullString
-			linkedTokenGUID        sql.NullString
-		)
-		err = rows.Scan(&tokenGUID, &ciphertextAuthToken, &ciphertextRefreshToken, &tokenExpiry, &disconnected, &authType, &metadata, &tokenUserGUID, &linkedTokenGUID)
-		if err != nil {
-			return nil, fmt.Errorf("Unable to scan CNSI records: %v", err)
-		}
-
-		log.Debug("Decrypting Auth Token")
-		plaintextAuthToken, err := crypto.DecryptToken(encryptionKey, ciphertextAuthToken)
-		if err != nil {
-			return make([]interfaces.TokenRecord, 0), err
-		}
-
-		log.Debug("Decrypting Refresh Token")
-		plaintextRefreshToken, err := crypto.DecryptToken(encryptionKey, ciphertextRefreshToken)
-		if err != nil {
-			return make([]interfaces.TokenRecord, 0), err
-		}
-
-		// Build a new TokenRecord based on the decrypted tokens
-		tr := new(interfaces.TokenRecord)
-		if tokenGUID.Valid {
-			tr.TokenGUID = tokenGUID.String
-		}
-		tr.AuthToken = plaintextAuthToken
-		tr.RefreshToken = plaintextRefreshToken
-		if tokenExpiry.Valid {
-			tr.TokenExpiry = tokenExpiry.Int64
-		}
-		tr.Disconnected = disconnected
-		tr.AuthType = authType
-		if metadata.Valid {
-			tr.Metadata = metadata.String
-		}
-		if tokenUserGUID.Valid {
-			tr.SystemShared = tokenUserGUID.String == SystemSharedUserGuid
-		}
-		if linkedTokenGUID.Valid {
-			tr.LinkedGUID = linkedTokenGUID.String
-		}
-
-		trs = append(trs, *tr)
-
-	}
-
-	// TODO: RC merge with find single
-	// TODO: RC Use?
-	// // If this token is linked - fetch that token and use it instead
-	// // Currently we don't recurse - we only support one level of linked token - you can't link to another linked token
-	// if linkedTokenGUID.Valid {
-	// 	if includeDisconnected {
-	// 		err = p.db.QueryRow(getToken, userGUID, linkedTokenGUID.String).Scan(&tokenGUID, &ciphertextAuthToken, &ciphertextRefreshToken, &tokenExpiry, &disconnected, &authType, &metadata, &tokenUserGUID, &linkedTokenGUID)
-	// 	} else {
-	// 		err = p.db.QueryRow(getTokenConnected, userGUID, linkedTokenGUID.String).Scan(&tokenGUID, &ciphertextAuthToken, &ciphertextRefreshToken, &tokenExpiry, &disconnected, &authType, &metadata, &tokenUserGUID, &linkedTokenGUID)
-	// 	}
-
-	// 	if err != nil {
-	// 		msg := "Unable to Find CNSI token: %v"
-	// 		if err == sql.ErrNoRows {
-	// 			log.Debugf(msg, err)
-	// 		} else {
-	// 			log.Errorf(msg, err)
-	// 		}
-	// 		return interfaces.TokenRecord{}, fmt.Errorf(msg, err)
-	// 	}
-	// }
-
-	// TODO: RC Finish off
-
-	return trs, nil
 }
 
 func (p *PgsqlTokenRepository) FindCNSIToken(cnsiGUID string, userGUID string, encryptionKey []byte) (interfaces.TokenRecord, error) {
@@ -775,17 +640,3 @@ func (p *PgsqlTokenRepository) UpdateTokenAuth(userGUID string, tr interfaces.To
 
 	return nil
 }
-
-// func (p *PgsqlTokenRepository) Overwrite(tr interfaces.TokenRecord, encryptionKey []byte) error {
-// 	log.Debug("Overwrite Token")
-
-// 	if _, err := p.FindCNSITokenIncludeDisconnected(endpointGuid, userGuid, encryptionKey); err != nil {
-// 		// Found, so update endpoint
-// 		// TODO: RC ALL STRINGS?
-// 		return p. Update(endpoint, encryptionKey)
-// 	} else {
-// 		// Not Found, create endpoint
-// 		return p.SaveCNSIToken(endpointGuid, userGuid, tr, encryptionKey)
-// 		// TODO: RC Q could actually be error
-// 	}
-// }
